inline useHomeDashboardData into HomeEnrollmentsDashboard
The hook is glue (composition of named helpers from the model layer)
with no domain logic of its own. A standalone hook file with no
isolated tests reads as "why no tests?" Inlining as a private function
in the component file keeps the data-vs-render split readable without
orphaning a file. Pure helpers in model/dashboardViewModel.ts continue
to carry the testable logic.

Add an open question to the plan doc: hook location and test strategy
for Phase 4/5 should be settled before those phases begin. Program and
contract dashboards may justify isolated hook tests on complexity
grounds even if home does not, and the language picker is a candidate
for its own reusable hook.

Deletes the now-empty hooks/ directory.
